requirements.txt
python-escpos[all]
python-dotenv
fastapi
uvicorn
python-multipart
